Add files via upload
We have made a Secured Data Encryption System using Python language and it is storing details in encrypted form of the user ensuring security to the user. The RSA technique used helps to prevent the third party from knowing the keys of the user. The Hash function prevents the Man in Middle attack. It can be implemented in the real world for personal uses.